refactor zaaktype en eigenschappen handlers
