+++ api/src/ActionHandler/ZaakTypeHandler.php
@@ -4,51 +4,39 @@ namespace App\ActionHandler;
 
 use App\Entity\ObjectEntity;
 use App\Exception\GatewayException;
+use App\Service\UhrZaakService;
 use Doctrine\ORM\EntityManagerInterface;
 use Psr\Container\ContainerInterface;
 
 class ZaakTypeHandler implements ActionHandlerInterface
 {
-    private EntityManagerInterface $entityManager;
+    private UhrZaakService $uhrZaakService;
 
     public function __construct(ContainerInterface $container)
     {
-        $entityManager = $container->get('doctrine.orm.entity_manager');
-        if ($entityManager instanceof EntityManagerInterface) {
-            $this->entityManager = $entityManager;
+        $zaakService = $container->get('uhrzaakservice');
+        if ($zaakService instanceof UhrZaakService) {
+            $this->uhrZaakService = $zaakService;
         } else {
             throw new GatewayException('The service container does not contain the required services for this handler');
         }
     }
 
-    public function getIdentifier(array $data, $configuration): string
-    {
-        $dotData = new \Adbar\Dot($data);
-
-        return $dotData->get($configuration['identifierPath']);
-    }
-
-    public function overridePath(string $value, string $path, array $data): array
-    {
-        $dotData = new \Adbar\Dot($data);
-        $dotData->set($path, $value);
-
-        return $dotData->jsonSerialize();
-    }
-
+    /**
+     * This function runs the zaak type plugin.
+     *
+     * @param array $data          The data from the call
+     * @param array $configuration The configuration of the action
+     *
+     * @throws \App\Exception\GatewayException
+     * @throws \Psr\Cache\CacheException
+     * @throws \Psr\Cache\InvalidArgumentException
+     * @throws \Respect\Validation\Exceptions\ComponentException
+     *
+     * @return array
+     */
     public function __run(array $data, array $configuration): array
     {
-        $identifier = $this->getIdentifier($data['request'], $configuration);
-        $entity = $this->entityManager->getRepository('App:Entity')->findOneBy(['id' => $configuration['entityId']]);
-        $objectEntities = $this->entityManager->getRepository('App:ObjectEntity')->findByEntity($entity, ['identificatie' => $identifier]);
-
-        if (count($objectEntities) > 0 && $objectEntities[0] instanceof ObjectEntity) {
-            $objectEntity = $objectEntities[0];
-
-            $url = $objectEntity->getValueByAttribute($objectEntity->getEntity()->getAttributeByName('url'))->getStringValue();
-            $data['request'] = $this->overridePath($url, $configuration['identifierPath'], $data['request']);
-        }
-
-        return $data;
+       return $this->uhrZaakService->zaakTypeHandler($data, $configuration);
     }
 }
